+++ Assets/OkapiKit/Scripts/Editor/ActionEditor.cs
@@ -64,6 +64,24 @@ namespace OkapiKit.Editor
             }
         }
 
+        public virtual void OnSceneGUI()
+        {
+            // Make sure to update the serializedObject to reflect the latest data
+            serializedObject.Update();
+
+            if (!propHasConditions.boolValue) return;
+
+            // Iterate through all elements in the propConditions array
+            for (int i = 0; i < propConditions.arraySize; i++)
+            {
+                SerializedProperty conditionElement = propConditions.GetArrayElementAtIndex(i);
+
+                // Render the property we want 
+                ConditionDrawer.OnSceneGUI(serializedObject, conditionElement);
+            }
+        }
+
+
         protected override Texture2D GetIcon()
         {
             var varTexture = GUIUtils.GetTexture("Action");