+++ ReleaseNotes.md
@@ -13,6 +13,7 @@
 - Added center path option on path object
 - Improved display and UI of path object
 - Tag count condition can now be limited to a specific range
+- Tag count with condition can now be rendered by in the scene view to be able to view the range
 
 ## V1.9.2